feat:(core) - Restructure OverviewComponent folder
