+++ src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { OverviewComponent } from './components/overview/overview/overview.component';
 import { TopicDetailsComponent } from './components/topic-details/topic-details/topic-details.component';
+import { OverviewComponent } from './components/overview/overview.component';
 
 const routes: Routes = [
   {
